+++ framework/include/executioners/Executioner.h
@@ -15,6 +15,7 @@
 #include "PostprocessorInterface.h"
 #include "Restartable.h"
 #include "PerfGraphInterface.h"
+#include "FEProblemSolve.h"
 #include "PicardSolve.h"
 
 // System includes
@@ -28,12 +29,6 @@ InputParameters validParams<Executioner>();
 
 /**
  * Executioners are objects that do the actual work of solving your problem.
- *
- * In general there are two "sets" of Executioners: Steady and Transient.
- *
- * The Difference is that a Steady Executioner usually only calls "solve()"
- * for the NonlinearSystem once... where Transient Executioners call solve()
- * multiple times... i.e. once per timestep.
  */
 class Executioner : public MooseObject,
                     public UserObjectInterface,
@@ -49,12 +44,12 @@ public:
    */
   Executioner(const InputParameters & parameters);
 
-  virtual ~Executioner();
+  virtual ~Executioner() {}
 
   /**
    * Initialize the executioner
    */
-  virtual void init();
+  virtual void init() {}
 
   /**
    * Pure virtual execute function MUST be overridden by children classes.
@@ -65,22 +60,22 @@ public:
   /**
    * Override this for actions that should take place before execution
    */
-  virtual void preExecute();
+  virtual void preExecute() {}
 
   /**
    * Override this for actions that should take place after execution
    */
-  virtual void postExecute();
+  virtual void postExecute() {}
 
   /**
-   * Override this for actions that should take place before execution
+   * Override this for actions that should take place before execution, called by PicardSolve
    */
-  virtual void preSolve();
+  virtual void preSolve() {}
 
   /**
-   * Override this for actions that should take place after execution
+   * Override this for actions that should take place after execution, called by PicardSolve
    */
-  virtual void postSolve();
+  virtual void postSolve() {}
 
   /**
    * Deprecated:
@@ -97,7 +92,12 @@ public:
    * This is an empty string for non-Transient executioners
    * @return A string of giving the TimeStepper name
    */
-  virtual std::string getTimeStepperName();
+  virtual std::string getTimeStepperName() { return std::string(); }
+
+  /**
+   * Reinitialize petsc output for proper linear/nonlinear iteration display
+   */
+  void initPetscOutput();
 
   /**
    * Can be used by subsclasses to call parentOutputPositionChanged()
@@ -110,11 +110,40 @@ public:
    */
   virtual bool lastSolveConverged() const = 0;
 
+  /// Return the underlining FEProblemSolve object.
+  FEProblemSolve & feProblemSolve() { return _feproblem_solve; }
+
   /// Return underlining PicardSolve object.
   PicardSolve & picardSolve() { return _picard_solve; }
 
-  /// Augmented Picard convergence check that can be overridden by derived executioners
+  /// Augmented FEProblem solve failure issued by executioners
+  virtual bool augmentedFEProblemSolveFail() const { return false; }
+
+  /// Augmented Picard convergence check that to be called by PicardSolve and can be overridden by derived executioners
   virtual bool augmentedPicardConvergenceCheck() const { return false; }
+
+  /**
+   * Set an exception.  Usually this should not be directly called - but should be called through
+   * the mooseException() macro.
+   *
+   * @param message The error message about the exception.
+   */
+  virtual void setException(const std::string & message);
+
+  /**
+   * Whether or not an exception has occurred.
+   */
+  virtual bool hasException() const { return _has_exception; }
+
+  /**
+   * Clear the exception flag.
+   */
+  virtual void clearException() { _has_exception = false; }
+
+  /**
+   * Retrieve the exception message.
+   */
+  virtual const std::string & exceptionMessage() const { return _exception_message; }
 
 protected:
   /**
@@ -129,17 +158,17 @@ protected:
 
   FEProblemBase & _fe_problem;
 
+  FEProblemSolve _feproblem_solve;
   PicardSolve _picard_solve;
 
-  /// Initial Residual Variables
-  Real _initial_residual_norm;
-  Real _old_initial_residual_norm;
+  /// Whether or not an exception has occurred
+  bool _has_exception;
+
+  /// The error message to go with an exception
+  std::string _exception_message;
 
   // Restart
   std::string _restart_file_base;
-
-  // Splitting
-  std::vector<std::string> _splitting;
 };
 
 #endif // EXECUTIONER_H
