+++ framework/include/executioners/SolveObject.h
@@ -17,6 +17,7 @@ class SolveObject;
 class Executioner;
 class FEProblemBase;
 class NonlinearSystemBase;
+class AuxiliarySystem;
 
 class SolveObject : public MooseObject, public PerfGraphInterface
 {
@@ -29,12 +30,25 @@ public:
    */
   virtual bool solve() = 0;
 
+  /// Set the inner solve object wrapped by this object.
+  virtual void setInnerSolve(SolveObject & solve) { _inner_solve = &solve; }
+
 protected:
   /// Executioner used to construct this
   Executioner & _executioner;
   /// Reference to FEProblem
   FEProblemBase & _problem;
+  /// Displaced problem
+  std::shared_ptr<DisplacedProblem> _displaced_problem;
+  /// Mesh
+  MooseMesh & _mesh;
+  /// Displaced mesh
+  MooseMesh * _displaced_mesh;
   /// Reference to nonlinear system base for faster access
   NonlinearSystemBase & _nl;
+  /// Reference to auxiliary system for faster access
+  AuxiliarySystem & _aux;
+  /// SolveObject wrapped by this solve object
+  SolveObject * _inner_solve;
 };
 #endif // SOLVEOBJECT_H
