+++ framework/include/problems/EigenProblem.h
@@ -28,10 +28,6 @@ class EigenProblem : public FEProblemBase
 public:
   EigenProblem(const InputParameters & parameters);
 
-  virtual void solve() override;
-
-  virtual bool converged() override;
-
   virtual unsigned int getNEigenPairsRequired() { return _n_eigen_pairs_required; }
   virtual void setNEigenPairsRequired(unsigned int n_eigen_pairs)
   {
