+++ framework/include/problems/ExternalProblem.h
@@ -31,13 +31,13 @@ public:
   /**
    * Solve is implemented to providing syncing to/from the "transfer" mesh.
    */
-  virtual void solve() override final;
+  virtual void solve();
 
   /**
    * External problems should provide an override for converged, not inherit the default from
    * FEProblem that provides a reasonable implementation.
    */
-  virtual bool converged() override = 0;
+  virtual bool converged() = 0;
 
   /**
    * New interface for solving an External problem. "solve()" is finalized here to provide
