+++ framework/src/executioners/Executioner.C
@@ -10,9 +10,9 @@
 // Moose includes
 #include "Executioner.h"
 
-#include "FEProblem.h"
 #include "MooseApp.h"
 #include "MooseMesh.h"
+#include "FEProblem.h"
 #include "NonlinearSystem.h"
 #include "SlepcSupport.h"
 
@@ -25,6 +25,7 @@ InputParameters
 validParams<Executioner>()
 {
   InputParameters params = validParams<MooseObject>();
+  params += validParams<FEProblemSolve>();
   params += validParams<PicardSolve>();
 
   params.addDeprecatedParam<FileNameNoExtension>(
@@ -37,64 +38,6 @@ validParams<Executioner>()
 
   params.addParamNamesToGroup("restart_file_base", "Restart");
 
-  params.addParam<std::vector<std::string>>("splitting",
-                                            "Top-level splitting defining a "
-                                            "hierarchical decomposition into "
-                                            "subsystems to help the solver.");
-
-  std::set<std::string> line_searches = {"contact", "default", "none", "basic"};
-#ifdef LIBMESH_HAVE_PETSC
-  std::set<std::string> petsc_line_searches = Moose::PetscSupport::getPetscValidLineSearches();
-  line_searches.insert(petsc_line_searches.begin(), petsc_line_searches.end());
-#endif // LIBMESH_HAVE_PETSC
-  std::string line_search_string = Moose::stringify(line_searches, " ");
-  MooseEnum line_search(line_search_string, "default");
-  std::string addtl_doc_str(" (Note: none = basic)");
-  params.addParam<MooseEnum>(
-      "line_search", line_search, "Specifies the line search type" + addtl_doc_str);
-  MooseEnum line_search_package("petsc moose", "petsc");
-  params.addParam<MooseEnum>("line_search_package",
-                             line_search_package,
-                             "The solver package to use to conduct the line-search");
-  params.addParam<unsigned>("contact_line_search_allowed_lambda_cuts",
-                            2,
-                            "The number of times lambda is allowed to be cut in half in the "
-                            "contact line search. We recommend this number be roughly bounded by 0 "
-                            "<= allowed_lambda_cuts <= 3");
-  params.addParam<Real>("contact_line_search_ltol",
-                        "The linear relative tolerance to be used while the contact state is "
-                        "changing between non-linear iterations. We recommend that this tolerance "
-                        "be looser than the standard linear tolerance");
-
-// Default Solver Behavior
-#ifdef LIBMESH_HAVE_PETSC
-  params += Moose::PetscSupport::getPetscValidParams();
-#endif // LIBMESH_HAVE_PETSC
-  params.addParam<Real>("l_tol", 1.0e-5, "Linear Tolerance");
-  params.addParam<Real>("l_abs_step_tol", -1, "Linear Absolute Step Tolerance");
-  params.addParam<unsigned int>("l_max_its", 10000, "Max Linear Iterations");
-  params.addParam<unsigned int>("nl_max_its", 50, "Max Nonlinear Iterations");
-  params.addParam<unsigned int>("nl_max_funcs", 10000, "Max Nonlinear solver function evaluations");
-  params.addParam<Real>("nl_abs_tol", 1.0e-50, "Nonlinear Absolute Tolerance");
-  params.addParam<Real>("nl_rel_tol", 1.0e-8, "Nonlinear Relative Tolerance");
-  params.addParam<Real>("nl_abs_step_tol", 1.0e-50, "Nonlinear Absolute step Tolerance");
-  params.addParam<Real>("nl_rel_step_tol", 1.0e-50, "Nonlinear Relative step Tolerance");
-  params.addParam<bool>(
-      "snesmf_reuse_base",
-      true,
-      "Specifies whether or not to reuse the base vector for matrix-free calculation");
-  params.addParam<bool>("no_fe_reinit", false, "Specifies whether or not to reinitialize FEs");
-  params.addParam<bool>("compute_initial_residual_before_preset_bcs",
-                        false,
-                        "Use the residual norm computed *before* PresetBCs are imposed in relative "
-                        "convergence check");
-
-  params.addParamNamesToGroup("l_tol l_abs_step_tol l_max_its nl_max_its nl_max_funcs "
-                              "nl_abs_tol nl_rel_tol nl_abs_step_tol nl_rel_step_tol "
-                              "compute_initial_residual_before_preset_bcs",
-                              "Solver");
-  params.addParamNamesToGroup("no_fe_reinit", "Advanced");
-
   return params;
 }
 
@@ -106,86 +49,12 @@ Executioner::Executioner(const InputParameters & parameters)
     PerfGraphInterface(this),
     _fe_problem(*getCheckedPointerParam<FEProblemBase *>(
         "_fe_problem_base", "This might happen if you don't have a mesh")),
+    _feproblem_solve(this),
     _picard_solve(this),
-    _initial_residual_norm(std::numeric_limits<Real>::max()),
-    _old_initial_residual_norm(std::numeric_limits<Real>::max()),
-    _restart_file_base(getParam<FileNameNoExtension>("restart_file_base")),
-    _splitting(getParam<std::vector<std::string>>("splitting"))
+    _has_exception(false),
+    _restart_file_base(getParam<FileNameNoExtension>("restart_file_base"))
 {
-  if (_pars.isParamSetByUser("line_search"))
-    _fe_problem.addLineSearch(_pars);
-
-// Extract and store PETSc related settings on FEProblemBase
-#ifdef LIBMESH_HAVE_PETSC
-  Moose::PetscSupport::storePetscOptions(_fe_problem, _pars);
-#endif // LIBMESH_HAVE_PETSC
-
-  // solver params
-  EquationSystems & es = _fe_problem.es();
-  es.parameters.set<Real>("linear solver tolerance") = getParam<Real>("l_tol");
-
-  es.parameters.set<Real>("linear solver absolute step tolerance") =
-      getParam<Real>("l_abs_step_tol");
-
-  es.parameters.set<unsigned int>("linear solver maximum iterations") =
-      getParam<unsigned int>("l_max_its");
-
-  es.parameters.set<unsigned int>("nonlinear solver maximum iterations") =
-      getParam<unsigned int>("nl_max_its");
-
-  es.parameters.set<unsigned int>("nonlinear solver maximum function evaluations") =
-      getParam<unsigned int>("nl_max_funcs");
-
-  es.parameters.set<Real>("nonlinear solver absolute residual tolerance") =
-      getParam<Real>("nl_abs_tol");
-
-  es.parameters.set<Real>("nonlinear solver relative residual tolerance") =
-      getParam<Real>("nl_rel_tol");
-
-  es.parameters.set<Real>("nonlinear solver absolute step tolerance") =
-      getParam<Real>("nl_abs_step_tol");
-
-  es.parameters.set<Real>("nonlinear solver relative step tolerance") =
-      getParam<Real>("nl_rel_step_tol");
-
-  _fe_problem.getNonlinearSystemBase()._compute_initial_residual_before_preset_bcs =
-      getParam<bool>("compute_initial_residual_before_preset_bcs");
-
-  _fe_problem.getNonlinearSystemBase()._l_abs_step_tol = getParam<Real>("l_abs_step_tol");
-
-  _fe_problem.setSNESMFReuseBase(getParam<bool>("snesmf_reuse_base"),
-                                 parameters.isParamSetByUser("snesmf_reuse_base"));
-
-  if (getParam<Real>("relaxation_factor") != 1.0)
-    // Store a copy of the previous solution here
-    _fe_problem.getNonlinearSystemBase().addVector("relax_previous", false, PARALLEL);
-}
-
-Executioner::~Executioner() {}
-
-void
-Executioner::init()
-{
-}
-
-void
-Executioner::preExecute()
-{
-}
-
-void
-Executioner::postExecute()
-{
-}
-
-void
-Executioner::preSolve()
-{
-}
-
-void
-Executioner::postSolve()
-{
+  _picard_solve.setInnerSolve(_feproblem_solve);
 }
 
 Problem &
@@ -201,10 +70,18 @@ Executioner::feProblem()
   return _fe_problem;
 }
 
-std::string
-Executioner::getTimeStepperName()
+void
+Executioner::initPetscOutput()
 {
-  return std::string();
+  _app.getOutputWarehouse().solveSetup();
+  Moose::PetscSupport::petscSetDefaults(_fe_problem);
+}
+
+void
+Executioner::setException(const std::string & message)
+{
+  _has_exception = true;
+  _exception_message = message;
 }
 
 void
