+++ framework/src/executioners/Steady.C
@@ -35,8 +35,6 @@ Steady::Steady(const InputParameters & parameters)
     _time(_problem.time()),
     _final_timer(registerTimedSection("final", 1))
 {
-  _problem.getNonlinearSystemBase().setDecomposition(_splitting);
-
   if (!_restart_file_base.empty())
     _problem.setRestartFile(_restart_file_base);
