+++ framework/src/postprocessors/Residual.C
@@ -12,6 +12,7 @@
 #include "FEProblem.h"
 #include "SubProblem.h"
 #include "NonlinearSystem.h"
+#include "FEProblemSolve.h"
 
 registerMooseObject("MooseApp", Residual);
 
@@ -39,16 +40,13 @@ Residual::getValue()
 {
   Real residual = 0.0;
   if (_residual_type == "FINAL")
-    residual = _subproblem.finalNonlinearResidual();
+    residual = _fe_problem.getFEProblemSolve().finalNonlinearResidual();
   else
   {
-    FEProblemBase * fe_problem = dynamic_cast<FEProblemBase *>(&_subproblem);
-    if (!fe_problem)
-      mooseError("Dynamic cast to FEProblemBase failed in Residual Postprocessor");
     if (_residual_type == "INITIAL_BEFORE_PRESET")
-      residual = fe_problem->getNonlinearSystemBase()._initial_residual_before_preset_bcs;
+      residual = _fe_problem.getNonlinearSystemBase()._initial_residual_before_preset_bcs;
     else if (_residual_type == "INITIAL_AFTER_PRESET")
-      residual = fe_problem->getNonlinearSystemBase()._initial_residual_after_preset_bcs;
+      residual = _fe_problem.getNonlinearSystemBase()._initial_residual_after_preset_bcs;
     else
       mooseError("Invalid residual_type option in Residual Postprocessor: ", _residual_type);
   }
