+++ framework/src/timeintegrators/LStableDirk3.C
@@ -95,7 +95,7 @@ LStableDirk3::solve()
     // This ensures that all the Output objects in the OutputWarehouse
     // have had solveSetup() called, and sets the default solver
     // parameters for PETSc.
-    _fe_problem.initPetscOutput();
+    _app.getExecutioner()->initPetscOutput();
 
     _console << "Stage " << _stage << "\n";
 
@@ -110,7 +110,7 @@ LStableDirk3::solve()
     _n_linear_iterations += getNumLinearIterationsLastSolve();
 
     // Abort time step immediately on stage failure - see TimeIntegrator doc page
-    if (!_fe_problem.converged())
+    if (!_fe_problem.getNonlinearSystemBase().converged())
       return;
   }
 }
