+++ framework/src/timeintegrators/LStableDirk4.C
@@ -90,7 +90,7 @@ LStableDirk4::solve()
     // This ensures that all the Output objects in the OutputWarehouse
     // have had solveSetup() called, and sets the default solver
     // parameters for PETSc.
-    _fe_problem.initPetscOutput();
+    _app.getExecutioner()->initPetscOutput();
 
     _console << "Stage " << _stage << "\n";
 
@@ -105,7 +105,7 @@ LStableDirk4::solve()
     _n_linear_iterations += getNumLinearIterationsLastSolve();
 
     // Abort time step immediately on stage failure - see TimeIntegrator doc page
-    if (!_fe_problem.converged())
+    if (!_fe_problem.getNonlinearSystemBase().converged())
       return;
   }
 }
